requirements.txt
altgraph==0.17.4
certifi==2021.10.8
chardet==4.0.0
dis3==0.1.3
future==1.0.0
idna==2.10
numpy==1.16.6
opencv-python==4.2.0.32
pefile==2019.4.18
Pillow==6.2.2
PyInstaller==3.6
pywin32-ctypes==0.2.0
requests==2.27.1
urllib3==1.26.18
